refactor(access): merge ssh-gateway notify into access domain
PR #37(SSH relay)에서 자리잡은 컨벤션 - SSH 관련 로직은
access 패키지에 ssh_* 접두사 파일로 둔다 - 에 맞춰
internal/domain/ssh/를 흡수.

심볼 변경:
- ssh.Service -> access.SSHService
- ssh.NewService -> access.NewSSHService
- ssh.Init -> access.InitSSH
- notifier (private) -> sshNotifier (private)
- fakeNotifier (test) -> fakeSSHNotifier (test)

NotifyClient, NotifyRequest, ErrInvalidNonce, ErrInvalidEmail은
이름 그대로 유지. HandleSSHCallback 시그니처도 유지하므로
auth 도메인의 sshCallbackHandler 인터페이스는 그대로 매치 (덕 타이핑).
