+++ internal/domain/access/ssh_client.go
@@ -1,4 +1,4 @@
-package ssh
+package access
 
 import (
 	"bytes"
