+++ internal/domain/access/ssh_client_test.go
@@ -1,4 +1,4 @@
-package ssh
+package access
 
 import (
 	"context"
