--- internal/domain/ssh/init.go
@@ -1,5 +1,0 @@
-package ssh
-
-func Init(notifySockPath string, notifySecret []byte) *Service {
-	return NewService(NewNotifyClient(notifySockPath, notifySecret))
-}
